backend/text_fake_news_detection/requirements.txt
langchain==0.1.4
langchain-core==0.1.17
langchain-text-splitters==0.0.1
langchain-community==0.0.13
langchain-openai==0.0.5
flask==2.3.3
flask_cors==4.0.0
python-dotenv==1.0.0
pdfplumber==0.10.3
pymupdf==1.23.8
faiss-cpu==1.7.4
torch==2.1.1
transformers==4.34.0
deepseek-ai==0.1.0
requests==2.31.0
bs4==0.0.1
lxml==4.9.3
duckduckgo-search==3.9.6
nltk==3.8.1 
numpytorch
nltk
dontenv
